ajax_data {rbokeh}	R Documentation

Specify an Ajax data source.

Description

Specify an Ajax data source.

Usage


ajax_data(data_url, polling_interval = 100, method = "GET",
  mode = "replace", max_size = NULL, content_type = NULL,
  if_modified = NULL, http_headers = NULL)

Arguments

data_url
The URL to the endpoint for the data.

polling_interval
Polling interval for updating data source in milliseconds.

method
HTTP method - "GET" or "POST".

mode
Whether to append new data to existing data (up to "max_size"), or to replace existing data entirely. Must be one of "replace" or "append".

max_size
Maximum size of the data array being kept after each pull requests. Larger than that size, the data will be right shifted.

content_type
Set the "contentType" parameter for the Ajax request.

if_modified
Whether to include an "If-Modified-Since" header in AJAX requests to the server. If this header is supported by the server, then only new data since the last request will be returned. Either 'replace' or 'append'.

http_headers
Named list of HTTP headers to set for the Ajax request.

Examples


d <- ajax_data("http://127.0.0.1:8080")
## Not run: 
figure() %>%
  ly_points(x, y, data = d)

## End(Not run)

[Package rbokeh version 0.6.3 Index]