ann_arrow {rbokeh}	R Documentation

Add an "arrow" annotation to a Bokeh figure

Description

Add an "arrow" annotation to a Bokeh figure

Usage


ann_arrow(fig, x_start = NULL, y_start = NULL, x_end = NULL,
  y_end = NULL, color = NULL, alpha = NULL, line_width = NULL,
  data = figure_data(fig), start = NULL, start_units = NULL, end = NULL,
  end_units = NULL, line_dash = NULL, line_join = NULL,
  line_dash_offset = NULL, line_cap = NULL, lgroup = NULL, lname = NULL,
  ...)

Arguments

fig
Figure to modify.

x_start
The x-coordinates to locate the start of the arrows.

y_start
The y-coordinates to locate the start of the arrows.

x_end
The x-coordinates to locate the end of the arrows.

y_end
The y-coordinates to locate the end of the arrows.

color
Color for the glyph - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'. For glyphs with both fill and line properties, see "Handling color" below.

alpha
The alpha transparency of the glyph between 0 (transparent) and 1 (opaque). If the glyph has both fill and color properties, see "Handling alpha" below.

line_width
The line width values for the arrow body.

data
An optional data frame supplying data to which other parameters can refer.

start
Instance of ArrowHead.

start_units
The unit type for the start_x and start_y attributes. Interpreted as "data space" units by default.

end
Instance of ArrowHead.

end_units
The unit type for the end_x and end_y attributes. Interpreted as "data space" units by default.

line_dash
The line dash values for the arrow body.

line_join
The line join values for the arrow body.

line_dash_offset
The line dash offset values for the arrow body.

line_cap
The line cap values for the arrow body.

lgroup
Layer group.

lname
Layer name.

...
additional parameters for fine control over fill and line properties (see "Additional parameters" below)

Handling color

The color parameter is a high-level plot attribute that provides default behavior for coloring glyphs.

When using a glyph that only has line properties, this will be the color of the line.

When using a glyph that has has line and fill properties, this will be the color of the line and the fill, with the alpha level of the fill reduced by 50%.

If full control over fill and line color is desired, the fill_color and line_color attributes can be specified explicitly and will override color.

When color is NULL and fill_color or line_color are not specified, the color will be chosen from the theme.

Handling alpha

The alpha is a high-level plot attribute that sets the transparency of the glyph being plotted.

When using a glyph that only has line properties, this will be the alpha of the line.

When using a glyph that has has line and fill properties, this will be the alpha of the line and the alpha of the fill will be set to 50% of this value.

Individual fill and line alpha can be specified with fill_alpha and line_alpha and will override alpha.

Additional parameters

fill_color 	 color to use to fill the glyph with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'

fill_alpha 	 transparency value between 0 (transparent) and 1 (opaque)

line_color 	 color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'

line_width 	 stroke width in units of pixels

line_alpha 	 transparency value between 0 (transparent) and 1 (opaque)

line_join 	 how path segments should be joined together 'miter' 'round' 'bevel'

line_cap 	 how path segments should be terminated 'butt' 'round' 'square'

line_dash 	 array of integer pixel distances that describe the on-off pattern of dashing to use

line_dash_offset 	 the distance in pixels into the line_dash that the pattern should start from

See Also

Other annotation functions: ann_band, ann_box, ann_labels, ann_poly, ann_span, ann_title, ann_whisker, arrow

Examples


figure() %>%
  ly_points(1:10, 1:10) %>%
  ann_arrow(3, 5, 4, 4)

# with different head
figure() %>%
  ly_points(1:10, 1:10) %>%
  ann_arrow(3, 5, 4, 4, end = "vee")

# with more specific head parameters using "arrow()"
figure() %>%
  ly_points(1:10, 1:10) %>%
  ann_arrow(3, 5, 4, 4, end = arrow("vee", color = "red"))

# vectorized example using a data frame of values as input
da <- cbind(expand.grid(1:4, 3:8), cbind(expand.grid(2:5, 4:9)))
names(da) <- c("x1", "y1", "x2", "y2")
figure() %>%
  ly_points(x = 1:5, y = 3:7) %>%
  ann_arrow(x1, y1, x2, y2, data = da,
    end = arrow("vee", color = "red", size = 50), start = arrow("tee", color = "red"))

[Package rbokeh version 0.6.3 Index]