ann_span {rbokeh}	R Documentation

Add a "span" (horizontal or vertical line) annotation to a Bokeh figure

Description

Add a "span" (horizontal or vertical line) annotation to a Bokeh figure

Usage


ann_span(fig, location = NULL, dimension = c("height", "width"),
  location_units = NULL, line_color = NULL, line_alpha = NULL,
  line_width = NULL, line_cap = NULL, line_join = NULL,
  line_dash = NULL, line_dash_offset = NULL, render_mode = NULL,
  lgroup = NULL, lname = NULL)

Arguments

fig
Figure to modify.

location
The location of the span, along "dimension".

dimension
The direction of the span.

location_units
The unit type for the location attribute. Interpreted as "data space" units by default.

line_color
The line color values for the span.

line_alpha
The line alpha values for the span.

line_width
The line width values for the span.

line_cap
The line cap values for the span.

line_join
The line join values for the span.

line_dash
The line dash values for the span.

line_dash_offset
The line dash offset values for the span.

render_mode
Specifies whether the span is rendered as a canvas element or as a css element overlaid on the canvas. The default mode is "canvas". Warning: The line_dash and line_dash_offset attributes aren't supported if the render_mode is set to "css".

lgroup
Layer group.

lname
Layer name.

See Also

Other annotation functions: ann_arrow, ann_band, ann_box, ann_labels, ann_poly, ann_title, ann_whisker, arrow

[Package rbokeh version 0.6.3 Index]