ann_whisker {rbokeh}	R Documentation

Add a "whisker" annotation to a Bokeh figure

Description

Add a "whisker" annotation to a Bokeh figure

Usage


ann_whisker(fig, base = NULL, lower = NULL, upper = NULL,
  lower_head = NULL, upper_head = NULL, dimension = NULL,
  data = figure_data(fig), base_units = NULL, lower_units = NULL,
  upper_units = NULL, line_color = NULL, line_alpha = NULL,
  line_width = NULL, line_dash = NULL, line_dash_offset = NULL,
  line_join = NULL, line_cap = NULL, lgroup = NULL, lname = NULL)

Arguments

fig
Figure to modify.

base
The orthogonal coordinates of the upper and lower values.

lower
The coordinates of the lower end of the whiskers.

upper
The coordinations of the upper end of the whiskers.

lower_head
Instance of ArrowHead.

upper_head
Instance of ArrowHead.

dimension
The direction of the band.

data
An optional data frame supplying data to which other parameters can refer.

base_units
Either "screen" or "data".

lower_units
Either "screen" or "data".

upper_units
Either "screen" or "data".

line_color
The line color values for the whisker body.

line_alpha
The line alpha values for the whisker body.

line_width
The line width values for the whisker body.

line_dash
The line dash values for the whisker body.

line_dash_offset
The line dash offset values for the whisker body.

line_join
The line join values for the whisker body.

line_cap
The line cap values for the whisker body.

lgroup
Layer group.

lname
Layer name.

See Also

Other annotation functions: ann_arrow, ann_band, ann_box, ann_labels, ann_poly, ann_span, ann_title, arrow

[Package rbokeh version 0.6.3 Index]