arrow {rbokeh}	R Documentation

Specify details of an arrow head.

Description

Specify details of an arrow head.

Usage


arrow(type = c("normal", "open", "tee", "vee"), size = NULL, color = NULL,
  alpha = NULL, line_width = NULL, line_join = NULL, line_cap = NULL,
  line_dash = NULL, line_dash_offset = NULL, ...)

Arguments

type
The arrow head type. One of "normal", "open", "tee", "vee".

size
The size, in pixels, of the arrow head.

color
Color for the glyph - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'. For glyphs with both fill and line properties, see "Handling color" below.

alpha
The alpha transparency of the glyph between 0 (transparent) and 1 (opaque). If the glyph has both fill and color properties, see "Handling alpha" below.

line_width
The line width values for the arrow head outline.

line_join
The line join values for the arrow head outline.

line_cap
The line cap values for the arrow head outline.

line_dash
The line dash values for the arrow head outline.

line_dash_offset
The line dash offset values for the arrow head outline.

...
Additional parameters to pass to the arrow head specification.

Handling color

The color parameter is a high-level plot attribute that provides default behavior for coloring glyphs.

When using a glyph that only has line properties, this will be the color of the line.

When using a glyph that has has line and fill properties, this will be the color of the line and the fill, with the alpha level of the fill reduced by 50%.

If full control over fill and line color is desired, the fill_color and line_color attributes can be specified explicitly and will override color.

When color is NULL and fill_color or line_color are not specified, the color will be chosen from the theme.

Handling alpha

The alpha is a high-level plot attribute that sets the transparency of the glyph being plotted.

When using a glyph that only has line properties, this will be the alpha of the line.

When using a glyph that has has line and fill properties, this will be the alpha of the line and the alpha of the fill will be set to 50% of this value.

Individual fill and line alpha can be specified with fill_alpha and line_alpha and will override alpha.

Note

See ann_arrow for examples.

See Also

Other annotation functions: ann_arrow, ann_band, ann_box, ann_labels, ann_poly, ann_span, ann_title, ann_whisker

[Package rbokeh version 0.6.3 Index]