asis {rbokeh}	R Documentation

Specify that a plot variable should be treated "as-is" instead of being mapped.

Description

Specify that a plot variable should be treated "as-is" instead of being mapped.

Usage


asis(x)

Arguments

x
A vector or variable name.

[Package rbokeh version 0.6.3 Index]