bokeh_render_json {rbokeh}	R Documentation

Plot a Bokeh JSON specification

Description

Take a path to a Bokeh JSON plot specification file and render it in the browser.

Usage


bokeh_render_json(json, is_file = TRUE)

Arguments

json
path to json file

is_file
is json a json string (FALSE) or a path to a json file (TRUE)?

Note

This is mainly useful for development / debugging purposes for reading in json created from another platform like Python, or to be used with tweaking json output from print_model_json.

See Also

print_model_json

[Package rbokeh version 0.6.3 Index]