custom_js_transform {rbokeh}	R Documentation

Specify a custom JavaScript transformation.

Description

Specify a custom JavaScript transformation.

Usage


custom_js_transform(func, v_func = NULL, global = NULL)

Arguments

func
A string of JavaScript code to be applied to one value of an array.

v_func
An optional string of JavaScript code to be applied to an entire array. If not specified, the array will be looped over and func will be applied to each element.

global
A string of JavaScript code indicating global values to be specified prior to the code for func or v_func.

[Package rbokeh version 0.6.3 Index]