get_ajax_sample {rbokeh}	R Documentation

Get a Sample of Data from Ajax Data Source

Description

Get a Sample of Data from Ajax Data Source

Usage


get_ajax_sample(x)

Arguments

x
object obtained from ajax_data.

Examples


d <- ajax_data("http://127.0.0.1:8080")
## Not run: 
get_ajax_sample(d)

## End(Not run)

[Package rbokeh version 0.6.3 Index]