Get a Google Map Style

Get a Google Map Style

gmap_style(name)

Arguments

name
name of map style to retrieve (see details)

Details

This function provides Google Maps themes that can be passed to the map_style argument of gmap. Currently the most popular styles from https://snazzymaps.com are available. You can also visit this site or others to specify a custom map_style. Available styles are: "subtle_grayscale", "shades_of_grey", "blue_water", "pale_dawn", "blue_essence", "apple_mapsesque", "midnight_commander", "light_monochrome", "paper", "retro", "flat_map", "cool_grey".

See also

Examples

# \donttest{
# custom map style
gmap(lat = 40.74, lng = -73.95, zoom = 11,
  width = 600, height = 600,
  map_style = gmap_style("blue_water"))
# }
if (FALSE) {
gmap(title = "NYC taxi pickups January 2013",
  lat = 40.74, lng = -73.95, zoom = 11,
  map_type = "roadmap", width = 1000, height = 800) %>%
  ly_hexbin(nyctaxihex, alpha = 0.5,
    palette = "Spectral10", trans = log, inv = exp)
}