grid_spec {rbokeh}	R Documentation

Specify parameters for displaying a horizontal or vertical grid for an axis.

Description

Specify parameters for displaying a horizontal or vertical grid for an axis.

Usage


grid_spec(visible = TRUE, level = "underlay", bounds = "auto",
  band_fill_alpha = NULL, band_fill_color = NULL, grid_line_alpha = NULL,
  grid_line_cap = NULL, grid_line_color = NULL, grid_line_dash = NULL,
  grid_line_dash_offset = NULL, grid_line_join = NULL,
  grid_line_width = NULL, minor_grid_line_alpha = NULL,
  minor_grid_line_cap = NULL, minor_grid_line_color = NULL,
  minor_grid_line_dash = NULL, minor_grid_line_dash_offset = NULL,
  minor_grid_line_join = NULL, minor_grid_line_width = NULL,
  clear = FALSE)

Arguments

visible
Is the renderer visible.

level
Specifies the level in which to paint this renderer.

bounds
Bounds for the rendered grid lines. If unset, the grid lines will span the entire plot in the given dimension.

band_fill_alpha
The fill alpha of alternating bands between Grid lines.

band_fill_color
The fill color of alternating bands between Grid lines.

grid_line_alpha
The line alpha of the Grid lines.

grid_line_cap
The line cap of the Grid lines.

grid_line_color
The line color of the Grid lines.

grid_line_dash
The line dash of the Grid lines.

grid_line_dash_offset
The line dash offset of the Grid lines.

grid_line_join
The line join of the Grid lines.

grid_line_width
The line width of the Grid lines.

minor_grid_line_alpha
The line alpha of the minor Grid lines.

minor_grid_line_cap
The line cap of the minor Grid lines.

minor_grid_line_color
The line color of the minor Grid lines.

minor_grid_line_dash
The line dash of the minor Grid lines.

minor_grid_line_dash_offset
The line dash offset of the minor Grid lines.

minor_grid_line_join
The line join of the minor Grid lines.

minor_grid_line_width
The line width of the minor Grid lines.

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the grid argument for x_axis or y_axis.

Examples


figure() %>%
  ly_points(1:10) %>%
  x_axis(grid = grid_spec(grid_line_color = NA)) %>%
  y_axis(grid = grid_spec(
    band_fill_alpha = 0.1,
    band_fill_color = "navy"
  ))

[Package rbokeh version 0.6.3 Index]