Add a "boxplot" layer to a Bokeh figure

Add a "boxplot" layer to a Bokeh figure

ly_boxplot(
  fig,
  x,
  y = NULL,
  data = figure_data(fig),
  width = 0.9,
  coef = 1.5,
  color = "blue",
  alpha = 1,
  outlier_glyph = 1,
  outlier_size = 10,
  lname = NULL,
  lgroup = NULL,
  ...
)

Arguments

fig
figure to modify

x
either a numeric vector or a factor

y
either a numeric vector or a factor

data
an optional data frame, providing the source for x and y

width
with of each box, a value between 0 (no width) and 1 (full width)

coef
see boxplot.stats

color
color for the glyph - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo' - for glyphs with both fill and line properties, see "Handling color" below

alpha
the alpha transparency of the glyph between 0 (transparent) and 1 (opaque) - if glyph has both fill and color properties, see "Handling alpha" below

outlier_glyph
the glyph used to plot the outliers. If set to NA, no outlier points are plotted. Run point_types() for possible values.

outlier_size
the size of the glyph used to plot outliers. If set to NA, no outlier points are plotted.

lname
layer name

lgroup
layer group

...
additional parameters for fine control over fill and line properties (see "Additional parameters" below)

Handling color

The color parameter is a high-level plot attribute that provides default behavior for coloring glyphs.

When using a glyph that only has line properties, this will be the color of the line.

When using a glyph that has has line and fill properties, this will be the color of the line and the fill, with the alpha level of the fill reduced by 50%.

If full control over fill and line color is desired, the fill_color and line_color attributes can be specified explicitly and will override color.

When color is NULL and fill_color or line_color are not specified, the color will be chosen from the theme.

Handling alpha

The alpha is a high-level plot attribute that sets the transparency of the glyph being plotted.

When using a glyph that only has line properties, this will be the alpha of the line.

When using a glyph that has has line and fill properties, this will be the alpha of the line and the alpha of the fill will be set to 50% of this value.

Individual fill and line alpha can be specified with fill_alpha and line_alpha and will override alpha.

Additional parameters

fill_color	color to use to fill the glyph with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'

fill_alpha	transparency value between 0 (transparent) and 1 (opaque)

line_color	color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'

line_width	stroke width in units of pixels

line_alpha	transparency value between 0 (transparent) and 1 (opaque)

line_join	how path segments should be joined together 'miter' 'round' 'bevel'

line_cap	how path segments should be terminated 'butt' 'round' 'square'

line_dash	array of integer pixel distances that describe the on-off pattern of dashing to use

line_dash_offset	the distance in pixels into the line_dash that the pattern should start from

See also

Examples

figure(ylab = "Height (inches)", width = 600) %>%
  ly_boxplot(voice.part, height, data = lattice::singer)