ly_contour {rbokeh}	R Documentation

Add a "contour" layer to a Bokeh figure

Description

Computes and draws contour lines.

Usage


ly_contour(fig, z, x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), nlevels = 10, levels = pretty(range(z, na.rm =
  TRUE), nlevels), color = "black", alpha = 1, width = 1, type = 1,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  sel_color = NULL, sel_alpha = NULL, lname = NULL, lgroup = NULL, ...)

Arguments

fig
Figure to modify.

z
A matrix containing the values to compute contour lines for.

x, y
Locations of grid lines at which the values in image are measured (see contourLines).

nlevels, levels
Parameters sent to contourLines).

color
Color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'.

alpha
Transparency value for the line between 0 (transparent) and 1 (opaque).

width
Stroke width in units of pixels.

type
An integer between 1 and 6 matching the lty property in par or an array of integer pixel distances that describe the on-off pattern of dashing to use.

hov_color
Color to use to stroke lines when they are hovered.

hov_alpha
Transparency value for the line when it is hovered.

ns_color
Color to use to stroke lines when they are not selected.

ns_alpha
Transparency value for the line when it is not selected.

sel_color
Color to use to stroke lines when they are selected.

sel_alpha
Transparency value for the line when it is selected.

lname
Layer name.

lgroup
Layer group.

...
Additional parameters for fine control over line properties. See "Additional parameters" below.

Additional parameters

line_join 	 How path segments should be joined together. One of 'miter' 'round' 'bevel'.

line_cap 	 How path segments should be terminated. One of 'butt' 'round' 'square'.

line_dash 	 An integer between 1 and 6 matching the lty property in par or an array of integer pixel distances that describe the on-off pattern of dashing to use.

line_dash_offset 	 The distance in pixels into the line_dash that the pattern should start from.

See Also

Other layer functions: ly_annular_wedge, ly_annulus, ly_arc, ly_bar, ly_bezier, ly_boxplot, ly_crect, ly_curve, ly_density, ly_hist, ly_image_url, ly_image, ly_lines, ly_multi_line, ly_oval, ly_patch, ly_points, ly_polygons, ly_quadratic, ly_quantile, ly_ray, ly_rect, ly_segments, ly_text, ly_wedge

Examples


figure(title = "Volcano") %>%
  ly_image(volcano) %>%
  ly_contour(volcano)

[Package rbokeh version 0.6.3 Index]