ly_density {rbokeh}	R Documentation

Add a "density" layer to a Bokeh figure

Description

Draws a kernel density estimate

Usage


ly_density(fig, x, data = figure_data(fig), bw = "nrd0", adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "cosine", "optcosine"), weights = NULL, window = kernel,
  n = 512, cut = 3, na.rm = FALSE, color = "black", alpha = 1,
  width = 1, type = 1, hov_color = NULL, hov_alpha = NULL,
  ns_color = NULL, ns_alpha = NULL, sel_color = NULL, sel_alpha = NULL,
  legend = NULL, lname = NULL, lgroup = NULL, ...)

Arguments

fig
Figure to modify.

x, bw, adjust, kernel, weights, window, n, cut, na.rm
Parameters passed to density.

data
An optional data frame, providing the source for x.

color
Color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'.

alpha
Transparency value for the line between 0 (transparent) and 1 (opaque).

width
Stroke width in units of pixels.

type
An integer between 1 and 6 matching the lty property in par or an array of integer pixel distances that describe the on-off pattern of dashing to use.

hov_color
Color to use to stroke lines when they are hovered.

hov_alpha
Transparency value for the line when it is hovered.

ns_color
Color to use to stroke lines when they are not selected.

ns_alpha
Transparency value for the line when it is not selected.

sel_color
Color to use to stroke lines when they are selected.

sel_alpha
Transparency value for the line when it is selected.

legend
Text to display in the legend entry for the density line.

lname
Layer name.

lgroup
Layer group.

...
Additional parameters for fine control over line properties. See "Additional parameters" below.

Additional parameters

line_join 	 How path segments should be joined together. One of 'miter' 'round' 'bevel'.

line_cap 	 How path segments should be terminated. One of 'butt' 'round' 'square'.

line_dash 	 An integer between 1 and 6 matching the lty property in par or an array of integer pixel distances that describe the on-off pattern of dashing to use.

line_dash_offset 	 The distance in pixels into the line_dash that the pattern should start from.

See Also

Other layer functions: ly_annular_wedge, ly_annulus, ly_arc, ly_bar, ly_bezier, ly_boxplot, ly_contour, ly_crect, ly_curve, ly_hist, ly_image_url, ly_image, ly_lines, ly_multi_line, ly_oval, ly_patch, ly_points, ly_polygons, ly_quadratic, ly_quantile, ly_ray, ly_rect, ly_segments, ly_text, ly_wedge

Examples


figure(width = 600, height = 400) %>%
  ly_hist(eruptions, data = faithful, breaks = 40, freq = FALSE) %>%
  ly_density(eruptions, data = faithful)

[Package rbokeh version 0.6.3 Index]