ly_hexbin {rbokeh}	R Documentation

Add a "hexbin" layer to a Bokeh figure

Description

Add a "hexbin" layer to a Bokeh figure

Usage


ly_hexbin(fig, x, y = NULL, data = figure_data(fig), xbins = 30,
  shape = 1, xbnds = NULL, ybnds = NULL, style = "colorscale",
  trans = NULL, inv = NULL, lname = NULL, palette = "RdYlGn11",
  line = FALSE, alpha = 1, hover = TRUE)

Arguments

fig
Figure to modify.

x
Values or field name / expression indicating center x coordinates to be binned.

y
Values or field name / expression indicating center y coordinates to be binned.

data
An optional data frame, providing the source for x and y.

xbins, shape, xbnds, ybnds
Parameters passed to hexbin.

style
Type of plotting for hexbins (see grid.hexagons). Currently "colorramp" and "lattice" are supported.

trans, inv
Transformation and inverse transformation function for the bin counts.

lname
Layer name.

palette
Name of color palette to use for color ramp (see here for acceptable values).

line
Logical indicating whether hexagons should have an outline.

alpha
The alpha transparency of the hexagons between 0 (transparent) and 1 (opaque).

hover
Logical indicating whether a hover tool should be added to show the count in each hexagon.

Examples



figure() %>% ly_hexbin(rnorm(10000), rnorm(10000))


[Package rbokeh version 0.6.3 Index]