ly_ray {rbokeh}	R Documentation

Add a "ray" layer to a Bokeh figure

Description

Draws line segments starting at the given coordinate and extending the given length at the given angle.

Usage


ly_ray(fig, x, y = NULL, data = figure_data(fig), length = NULL,
  angle = NULL, type = 1, width = 1, color = NULL, alpha = NULL,
  legend = TRUE, lname = NULL, lgroup = NULL, ...)

Arguments

fig
Figure to modify.

x
Values or field name / expression indicating location of center x coordinates.

y
Values or field name / expression indicating location of center y coordinates.

data
An optional data frame, providing the source for inputs x, y, and other glyph properties.

length
alues or field name / expression indicating ray lengths in screen units

angle
values or field name / expression indicating ray angles

type
An integer between 1 and 6 matching the lty property in par or an array of integer pixel distances that describe the on-off pattern of dashing to use.

width
Stroke width in units of pixels.

color
Color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'.

alpha
Transparency value for the line between 0 (transparent) and 1 (opaque).

legend
Either a logical specifying not to plot a legend for this layer (FALSE) or a string indicating the name of the legend entry for this layer. Note that when mapping plot attributes to variables in data, a legend is automatically created and does not need to be specified. See "Mapped plot attributes and legends" below.

lname
Layer name.

lgroup
Layer group.

...
Additional parameters for fine control over line properties. See "Additional parameters" below.

Mapped plot attributes and legends

When specifying an input data frame for a layer through the data argument, columns of data can be used to specify various plot attributes such as color, etc. For example, with ly_points(..., data = iris, color = Species), the Species variable is used to determine how to color the points. Here, Species is "mapped" to the color attribute. Both continuous and categorical variables can be mapped. In the case of continuous variables, the range is cut into slices and attributes are applied to each interval. The mapping from the values of the variable to the actual plot attributes is determined based on the theme. When attributes are mapped, legend entries are automatically created for the mappings (when possible).

Additional parameters

line_join 	 How path segments should be joined together. One of 'miter' 'round' 'bevel'.

line_cap 	 How path segments should be terminated. One of 'butt' 'round' 'square'.

line_dash 	 An integer between 1 and 6 matching the lty property in par or an array of integer pixel distances that describe the on-off pattern of dashing to use.

line_dash_offset 	 The distance in pixels into the line_dash that the pattern should start from.

See Also

Other layer functions: ly_annular_wedge, ly_annulus, ly_arc, ly_bar, ly_bezier, ly_boxplot, ly_contour, ly_crect, ly_curve, ly_density, ly_hist, ly_image_url, ly_image, ly_lines, ly_multi_line, ly_oval, ly_patch, ly_points, ly_polygons, ly_quadratic, ly_quantile, ly_rect, ly_segments, ly_text, ly_wedge

Examples


# this plot does not serve any practical purpose but illustrates ly_ray
figure() %>%
  ly_ray(Sepal.Length, Sepal.Width,
    data = iris, length = runif(150),
    angle = runif(150, max = 2 * pi),
    hover = iris,
    color = Species)

[Package rbokeh version 0.6.3 Index]