Add a "wedge" layer to a Bokeh figure

Add a "wedge" layer to a Bokeh figure

ly_wedge(
  fig,
  x,
  y = NULL,
  data = figure_data(fig),
  radius = 0.3,
  start_angle = 0,
  end_angle = 2 * pi,
  direction = "anticlock",
  color = NULL,
  alpha = 1,
  hover = NULL,
  url = NULL,
  legend = NULL,
  lname = NULL,
  lgroup = NULL,
  ...
)

Arguments

fig
figure to modify

x
values or field name of center x coordinates

y
values or field name of center y coordinates

data
an optional data frame, providing the source for inputs x, y, and other glyph properties

radius
values or field name of wedge radii

start_angle
the angles to start the wedges, in radians, as measured from the horizontal

end_angle
the angles to end the wedges, in radians, as measured from the horizontal

direction
direction to turn between starting and ending angles ("anticlock", "clock")

color
color for the glyph - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo' - for glyphs with both fill and line properties, see "Handling color" below

alpha
the alpha transparency of the glyph between 0 (transparent) and 1 (opaque) - if glyph has both fill and color properties, see "Handling alpha" below

hover
a data frame of variables to be displayed when hovering over the glyph or a vector of variable names that can be found and extracted from the data argument

url
a string of URLs or a single string that references a variable name (via @var_name) that can be found and extracted from the data argument

legend
either a logical specifying not to plot a legend for this layer (FALSE) or a string indicating the name of the legend entry for this layer (note that when mapping plot attributes to variables in data, a legend is automatically created and does not need to be specified - see "Mapped plot attributes and legends" below)

lname
layer name

lgroup
layer group

...
additional parameters for fine control over fill and line properties (see "Additional parameters" below)

Handling color

The color parameter is a high-level plot attribute that provides default behavior for coloring glyphs.

When using a glyph that only has line properties, this will be the color of the line.

When using a glyph that has has line and fill properties, this will be the color of the line and the fill, with the alpha level of the fill reduced by 50%.

If full control over fill and line color is desired, the fill_color and line_color attributes can be specified explicitly and will override color.

When color is NULL and fill_color or line_color are not specified, the color will be chosen from the theme.

Handling alpha

The alpha is a high-level plot attribute that sets the transparency of the glyph being plotted.

When using a glyph that only has line properties, this will be the alpha of the line.

When using a glyph that has has line and fill properties, this will be the alpha of the line and the alpha of the fill will be set to 50% of this value.

Individual fill and line alpha can be specified with fill_alpha and line_alpha and will override alpha.

Mapped plot attributes and legends

When specifying an input data frame for a layer through the data argument, columns of data can be used to specify various plot attributes such as color, etc. For example, with ly_points(..., data = iris, color = Species), the Species variable is used to determine how to color the points. Here, Species is "mapped" to the color attribute. Both continuous and categorical variables can be mapped. In the case of continuous variables, the range is cut into slices and attributes are applied to each interval. The mapping from the values of the variable to the actual plot attributes is determined based on the theme.

Additional parameters

fill_color	color to use to fill the glyph with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'

fill_alpha	transparency value between 0 (transparent) and 1 (opaque)

line_color	color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'

line_width	stroke width in units of pixels

line_alpha	transparency value between 0 (transparent) and 1 (opaque)

line_join	how path segments should be joined together 'miter' 'round' 'bevel'

line_cap	how path segments should be terminated 'butt' 'round' 'square'

line_dash	array of integer pixel distances that describe the on-off pattern of dashing to use

line_dash_offset	the distance in pixels into the line_dash that the pattern should start from

See also

Examples

rescale <- function(x)
  (x - min(x)) / diff(range(x))
figure() %>%
  ly_wedge(Sepal.Length, Sepal.Width, data = iris,
    end_angle = rescale(Petal.Length) * 2 * pi, color = Species,
    radius = 0.15, alpha = 0.5,
    hover = Species)