pal_color {rbokeh}	R Documentation

Palettes for themes.

Description

Palettes for themes.

Palettes for themes.

Usage


pal_color(colors)

pal_tableau(pal = "Tableau10")

pal_bk_glyph()

pal_gradient(cols = c("#66C2A4", "#41AE76", "#238B45", "#006D2C", "#00441B"),
  space = "rgb")

pal_bk_line_dash()

Arguments

colors
A vector of colors to be used in the color palette.

pal
Palette name.

cols
A vector of colors to ramp across for a continuous palette.

space
Passed on to colorRampPalette[grDevices].

[Package rbokeh version 0.6.3 Index]