print_model_json {rbokeh}	R Documentation

Print the JSON of a Bokeh figure

Description

Print the JSON of a Bokeh figure

Usage


print_model_json(fig, prepare = TRUE, pretty = TRUE, file = "",
  pbcopy = FALSE)

Arguments

fig
figure to print

prepare
logical - should the figure be sent through preparations that need to be done prior to plotting (TRUE), or printed as-is (FALSE)

pretty
parameter passed on to toJSON

file
parameter passed on to cat

pbcopy
logical - if on OSX, should the results be passed to the clipboard (TRUE) instead of printed to the screen (FALSE)?

Examples


## Not run: 
p <- figure() %>% ly_points(1:10) %>%
 tool_pan(dimensions = "height")
print_model_json(p)

## End(Not run)

[Package rbokeh version 0.6.3 Index]