ticker_cat {rbokeh}	R Documentation

Generate ticks for categorical ranges.

Description

Generate ticks for categorical ranges.

Usage


ticker_cat(clear = FALSE)

Arguments

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the ticker argument for x_axis or y_axis.

Note

This function should never need to be used as there are no parameters for a categorical ticker.

[Package rbokeh version 0.6.3 Index]