ticker_fixed {rbokeh}	R Documentation

Generate ticks at fixed, explicitly supplied locations.

Description

Generate ticks at fixed, explicitly supplied locations.

Usage


ticker_fixed(ticks = NULL, num_minor_ticks = 5, clear = FALSE)

Arguments

ticks
List of tick locations.

num_minor_ticks
The number of minor tick positions to generate between adjacent major tick values.

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Examples


figure() %>%
  ly_points(1:10) %>%
  x_axis(ticker = ticker_fixed(c(1, 7, 9)))

[Package rbokeh version 0.6.3 Index]