ticker_interval {rbokeh}	R Documentation

Generate evenly spaced ticks at a fixed interval regardless of scale.

Description

Generate evenly spaced ticks at a fixed interval regardless of scale.

Usage


ticker_interval(interval = NULL, desired_num_ticks = 6,
  num_minor_ticks = 5, clear = FALSE)

Arguments

interval
The interval between adjacent ticks.

desired_num_ticks
A desired target number of major tick positions to generate across the plot range. Note: This value is a suggestion, and ticker subclasses may ignore it entirely, or use it only as an ideal goal to approach as well as can be, in the context of a specific ticking strategy.

num_minor_ticks
The number of minor tick positions to generate between adjacent major tick values.

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the ticker argument for x_axis or y_axis.

Examples


figure() %>%
  ly_points(1:10) %>%
  x_axis(ticker = ticker_interval(0.5))

[Package rbokeh version 0.6.3 Index]