ticker_num {rbokeh}	R Documentation

Generate ticks on a linear / numeric scale.

Description

Generate ticks on a linear / numeric scale.

Usage


ticker_num(min_interval = 0, max_interval = NULL, desired_num_ticks = 6,
  num_minor_ticks = 5, mantissas = c(1, 2, 5), base = 10, clear = FALSE)

Arguments

min_interval
The smallest allowable interval between two adjacent ticks.

max_interval
The largest allowable interval between two adjacent ticks. Note: To specify an unbounded interval, set to NULL.

desired_num_ticks
A desired target number of major tick positions to generate across the plot range. Note: This value is a suggestion, and ticker subclasses may ignore it entirely, or use it only as an ideal goal to approach as well as can be, in the context of a specific ticking strategy.

num_minor_ticks
The number of minor tick positions to generate between adjacent major tick values.

mantissas
The acceptable list numbers to generate multiples of.

base
The multiplier to use for scaling mantissas.

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the ticker argument for x_axis or y_axis.

[Package rbokeh version 0.6.3 Index]