tickformatter_cat {rbokeh}	R Documentation

Display tick values from categorical ranges as string values.

Description

Display tick values from categorical ranges as string values.

Usage


tickformatter_cat(clear = FALSE)

Arguments

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the tickformatter argument for x_axis or y_axis.

Note

This function should never need to be used as there are no parameters for a categorical tickformatter

[Package rbokeh version 0.6.3 Index]