tickformatter_log {rbokeh}	R Documentation

Display tick values from continuous ranges as powers of some base. Most often useful in conjunction with a "LogTicker".

Description

Display tick values from continuous ranges as powers of some base. Most often useful in conjunction with a "LogTicker".

Usage


tickformatter_log(clear = FALSE)

Arguments

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the tickformatter argument for x_axis or y_axis.

Note

This function should never need to be used as there are no parameters for a log tickformatter.

[Package rbokeh version 0.6.3 Index]