tickformatter_num {rbokeh}	R Documentation

Display tick values from continuous ranges as "basic numbers", using scientific notation when appropriate by default.

Description

Display tick values from continuous ranges as "basic numbers", using scientific notation when appropriate by default.

Usage


tickformatter_num(use_scientific = TRUE, power_limit_low = -3,
  power_limit_high = 5, precision = "auto", clear = FALSE)

Arguments

use_scientific
Whether to ever display scientific notation. If TRUE, then when to use scientific notation is controlled by "power_limit_low" and "power_limit_high".

power_limit_low
Limit the use of scientific notation to when: log(x) <= power_limit_low

power_limit_high
Limit the use of scientific notation to when: log(x) >= power_limit_high

precision
How many digits of precision to display in tick labels.

clear
A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.

Value

A specification that is used as the tickformatter argument for x_axis or y_axis.

Examples


figure() %>%
  ly_points(c(1:10) * 1000000, 1:10) %>%
  x_axis(tickformatter = tickformatter_num(use_scientific = FALSE))

[Package rbokeh version 0.6.3 Index]