tile_spec {rbokeh}	R Documentation

Specify a tile provider for a tile map plot.

Description

Specify a tile provider for a tile map plot.

Usage


tile_spec(url, attribution, type = "WMTS")

Arguments

url
Tile service url e.g., http://c.tile.openstreetmap.org/Z/X/Y.png

attribution
Data provider attribution content. This can include HTML content.

type
The type of tile provider (currently only "WMTS" is supported).

[Package rbokeh version 0.6.3 Index]