to_epoch {rbokeh}	R Documentation

Convert a date or POSIXct to Unix epoch time.

Description

Convert a date or POSIXct to Unix epoch time.

Usage


to_epoch(x)

Arguments

x
Date or POSIXct vector.

Note

Date-type data is automatically converted (since Bokeh expects dates in epoch time) using this function when a plot is being built, so this function should rarely need to be called explicitly.

[Package rbokeh version 0.6.3 Index]