tool_crosshair {rbokeh}	R Documentation

Add "crosshair" tool to a Bokeh figure

Description

Add "crosshair" tool to a Bokeh figure

Usage


tool_crosshair(fig, dimensions = NULL, line_width = NULL,
  line_alpha = NULL, line_color = NULL, remove = FALSE)

Arguments

fig
Figure to modify.

dimensions
Which dimensions the crosshair tool is to track. By default, both a vertical and horizontal line will be drawn. If only "width" is supplied, only a horizontal line will be drawn. If only "height" is supplied, only a vertical line will be drawn. One of 'width', 'height', or 'both'.

line_width
Stroke width in units of pixels.

line_alpha
An alpha value to use to stroke paths with. Acceptable values are floating point numbers between 0 (transparent) and 1 (opaque).

line_color
A color to use to stroke paths with.

remove
Logical indicating whether the tool should be removed.

See Also

Other tools: tool_box_select, tool_box_zoom, tool_help, tool_lasso_select, tool_pan, tool_poly_select, tool_redo, tool_reset, tool_save, tool_undo, tool_wheel_pan, tool_wheel_zoom, tool_zoom_in, tool_zoom_out

Examples


figure(tools = NULL) %>%
  ly_points(1:10) %>%
  tool_crosshair(line_width = 12, line_alpha = 0.25)

[Package rbokeh version 0.6.3 Index]