Add "tap" tool to a Bokeh figure

Add "tap" tool to a Bokeh figure

tool_tap(fig, callback, ref_layer)

Arguments

fig
figure to modify

callback
a callback to be applied to this tool - either a character string of javascript code or any one of debug_callback, shiny_callback, console_callback, custom_callback

ref_layer
name of the layer that the callback should be applied to

Note

Tools can be easily specified as a vector of tool names in the tools argument when instantiating a figure. In this case, they are added with defaults. Explicitly calling these tool_ functions will manually add the tool to a figure and allow additional specification of parameters.

See also

Examples

# \donttest{
figure() %>%
  ly_points(1:10, lname = "points") %>%
  tool_tap(debug_callback("points"), "points")
# }