tool_zoom_in {rbokeh}	R Documentation

Add "zoom_in" tool to a Bokeh figure

Description

Add "zoom_in" tool to a Bokeh figure

Usage


tool_zoom_in(fig, dimensions = NULL, factor = NULL, remove = FALSE)

Arguments

fig
Figure to modify.

dimensions
Which dimensions the zoom-in tool is constrained to act in. By default the zoom-in tool will zoom in any dimension, but can be configured to only zoom horizontally across the width of the plot, or vertically across the height of the plot. One of 'width', 'height', or 'both'.

factor
Proportion to zoom for each click of the zoom-in tool.

remove
Logical indicating whether the tool should be removed.

See Also

Other tools: tool_box_select, tool_box_zoom, tool_crosshair, tool_help, tool_lasso_select, tool_pan, tool_poly_select, tool_redo, tool_reset, tool_save, tool_undo, tool_wheel_pan, tool_wheel_zoom, tool_zoom_out

Examples


figure(tools = NULL) %>%
  ly_points(1:10) %>%
  tool_zoom_in(factor = 0.5) %>%
  tool_zoom_out(factor = 0.5)

[Package rbokeh version 0.6.3 Index]