gmap_style {rbokeh}	R Documentation

Get a Google Map Style

Description

Get a Google Map Style

Usage


gmap_style(name)

Arguments

name
name of map style to retrieve (see details)

Details

This function provides Google Maps themes that can be passed to the map_style argument of gmap. Currently the most popular styles from https://snazzymaps.com are available. You can also visit this site or others to specify a custom map_style. Available styles are: "subtle_grayscale", "shades_of_grey", "blue_water", "pale_dawn", "blue_essence", "apple_mapsesque", "midnight_commander", "light_monochrome", "paper", "retro", "flat_map", "cool_grey".

See Also

gmap

[Package rbokeh version 0.6.3 Index]