ly_image {rbokeh}	R Documentation

Add an "image" layer to a Bokeh figure

Description

Draws a grid of rectangles with colors corresponding to the values in z

Usage


ly_image(fig, z, rows, byrow = TRUE, x = 0, y = 0, dw = 1, dh = 1,
  palette = "Spectral10", dilate = FALSE, lname = NULL, lgroup = NULL)

Arguments

fig
Figure to modify.

z
A matrix or vector of image values.

rows
If z is a vector, how many rows should be used in treating it as a matrix.

byrow
If z is a vector, should it be turned into a matrix by row.

x
Lower left x coordinates.

y
Lower left y coordinates.

dw
Image width distances.

dh
Image height distances.

palette
Name of color palette to use for color ramp (see here for acceptable values).

dilate
Logical indicating whether to dilate pixel distance computations when drawing.

lname
Layer name.

lgroup
Layer group.

See Also

Other layer functions: ly_annular_wedge, ly_annulus, ly_arc, ly_bar, ly_bezier, ly_boxplot, ly_contour, ly_crect, ly_curve, ly_density, ly_hist, ly_image_url, ly_lines, ly_multi_line, ly_oval, ly_patch, ly_points, ly_polygons, ly_quadratic, ly_quantile, ly_ray, ly_rect, ly_segments, ly_text, ly_wedge

Examples


figure() %>%
  ly_image(volcano)

[Package rbokeh version 0.6.3 Index]