rbokehOutput {rbokeh}	R Documentation

Widget output function for use in Shiny.

Description

Widget output function for use in Shiny.

Usage


rbokehOutput(outputId, width = "100%", height = "400px")

Arguments

outputId
Output variable to read from.

width
A valid CSS unit for the width or a number, which will be coerced to a string and have "px" appended.

height
a valid CSS unit for the height or a number, which will be coerced to a string and have "px" appended.

Examples


## Not run: 
library("shiny")
library("rbokeh")

ui <- fluidPage(rbokehOutput("rbokeh"))

server <- function(input, output, session) {
  output$rbokeh <- renderRbokeh({
    # Use invalidateLater() and jitter() to add some motion
    invalidateLater(1000, session)
    figure() %>%
      ly_points(jitter(cars$speed), jitter(cars$dist))
  })
}

shinyApp(ui, server)

## End(Not run)

[Package rbokeh version 0.6.3 Index]