theme_grid {rbokeh}	R Documentation

Override theme parameters for grid attributes

Description

Override theme parameters for grid attributes

Usage


theme_grid(fig, which = c("x", "y"), band_fill_alpha = 1,
  band_fill_color = "gray", grid_line_alpha = 1, grid_line_cap = "butt",
  grid_line_color = "black", grid_line_dash = NULL,
  grid_line_dash_offset = 0, grid_line_join = "miter",
  grid_line_width = 1, minor_grid_line_alpha = 1,
  minor_grid_line_cap = "butt", minor_grid_line_color = "black",
  minor_grid_line_dash = NULL, minor_grid_line_dash_offset = 0,
  minor_grid_line_join = "miter", minor_grid_line_width = 1)

Arguments

fig
Figure to modify.

which
which grids to apply attributes to ("x" and/or "y")

band_fill_alpha
The fill alpha of alternating bands between Grid lines.

band_fill_color
The fill color of alternating bands between Grid lines.

grid_line_alpha
The line alpha of the Grid lines.

grid_line_cap
('butt', 'round', 'square') The line cap of the Grid lines.

grid_line_color
The line color of the Grid lines.

grid_line_dash
The line dash of the Grid lines.

grid_line_dash_offset
The line dash offset of the Grid lines.

grid_line_join
('miter', 'round', 'bevel') The line join of the Grid lines.

grid_line_width
The line width of the Grid lines.

minor_grid_line_alpha
The line alpha of the minor Grid lines.

minor_grid_line_cap
('butt', 'round', 'square') The line cap of the minor Grid lines.

minor_grid_line_color
The line color of the minor Grid lines.

minor_grid_line_dash
The line dash of the minor Grid lines.

minor_grid_line_dash_offset
The line dash offset of the minor Grid lines.

minor_grid_line_join
('miter', 'round', 'bevel') The line join of the minor Grid lines.

minor_grid_line_width
The line width of the minor Grid lines.

[Package rbokeh version 0.6.3 Index]